requirements.txt
pip~=24.2
matplotlib
music21
pillow
